Functions

Functions:

Converting a Vector of MOPSolution or BOPSolution into an array

Modof.wrap_sols_into_array — Function.

wrap_sols_into_array(sols::Vector{MOPSolution})

Convert a vector of MOPSolution into an Array{Float64,2}

wrap_sols_into_array(sols::Vector{BOPSolution})

Convert a vector of BOPSolution into an Array{Float64,2}

Computing objective function values

Modof.compute_objective_function_value! — Function.

compute_objective_function_value!{T<:MOOInstance, S<:MOOSolution}(solution::S, instance::T)

Compute objective function values of a MOOSolution of a Multi Objective Optimization Instance

compute_objective_function_value!{T<:MOOInstance, S<:MOOSolution}(solutions::Vector{S}, instance::T)

Compute objective function values of a vector{MOOSolution} of a Multi Objective Optimization Instance

compute_objective_function_value!{T<:BOOInstance, S<:BOOSolution}(solution::S, instance::T)

Compute objective function values of a BOOSolution of a Bi Objective Optimization Instance

compute_objective_function_value!{T<:BOOInstance, S<:BOOSolution}(solutions::Vector{S}, instance::T)

Compute objective function values of a vector{BOOSolution} of a Bi Objective Optimization Instance

Checking feasibility

Modof.check_feasibility — Function.

check_feasibility(solutions::Union{Vector{MOPSolution}, Vector{BOPSolution}}, instance::Union{MOBPInstance, MOIPInstance, MOMBLPInstance, MOMILPInstance, BOBPInstance, BOIPInstance, BOMBLPInstance, BOMILPInstance})

Check feasibility of a vector of biobjective or a vector of multiobjective optimization solutions and only returns the vector of feasible solutions

Selecting, sorting, writing and normalizing a nondominated frontier

Modof.check_dominance — Function.

check_dominance{T<:Number}(point1::Vector{T}, point2::Vector{T})

Checks whether points point1 and point2 dominate one another.

Modof.select_non_dom_sols — Function.

select_non_dom_sols{T<:Number}(solutions::Array{T,2})

Select nondominated points for both Biobjective and Multiobjective Programs.

select_non_dom_sols(solutions::Vector{MOPSolution})

Select nondominated solutions among solutions for Multiobjective Programs.

select_non_dom_sols(solutions::Vector{BOPSolution})

Select nondominated solutions among solutions for Biobjective Programs.

Modof.select_unique_sols — Function.

select_unique_sols{T<:Number}(solutions::Array{T, 2})

Select unique nondominated points for both Biobjective and Multiobjective Programs.

select_unique_sols(solutions::Vector{MOPSolution})

Select unique nondominated solutions among solutions for Multiobjective Programs.

select_non_dom_sols(solutions::Vector{BOPSolution})

Select unique nondominated solutions among solutions for Biobjective Programs.

Modof.sort_non_dom_sols — Function.

sort_non_dom_sols{T<:Number}(solutions::Array{T,2}, index::Int64=1)

Sort nondominated solutions for both Biobjective and Multiobjective Programs based on index objective.

sort_non_dom_sols(solutions::Vector{MOPSolution}, index::Int64=1)

Sort nondominated solutions for Multiobjective Programs based on index objective.

sort_non_dom_sols(solutions::Vector{BOPSolution}, index::Int64=1)

Sort nondominated solutions for Biobjective Programs based on index objective.

Modof.select_and_sort_non_dom_sols — Function.

select_and_sort_non_dom_sols{T<:Number}(solutions::Array{T,2}, index::Int64=1)

Select and Sort nondominated solutions for both Biobjective and Multiobjective Programs.

select_and_sort_non_dom_sols(solutions::Vector{MOPSolution}, index::Int64=1)

Select and sort nondominated solutions for Multiobjective Programs based on index objective.

select_and_sort_non_dom_sols(solutions::Vector{BOPSolution}, index::Int64=1)

Select and Sort nondominated solutions for Biobjective Programs based on index objective.

Modof.write_nondominated_frontier — Function.

write_nondominated_frontier(non_dom_sols::Union{Vector{MOPSolution}, Vector{BOPSolution}}, filename::String)

Write nondominated frontier of a Vector{BOPSolution} or a Vector{MOPSolution} and save it to filename.

Modof.write_nondominated_sols — Function.

write_nondominated_sols(non_dom_sols::Union{Vector{MOPSolution}, Vector{BOPSolution}}, filename::String)

Write nondominated solutions of a Vector{BOPSolution} or a Vector{MOPSolution} and save it to filename.

Computing ideal and nadir points of a nondominated frontier

Modof.compute_ideal_point — Function.

compute_ideal_point{T<:Number}(non_dom_sols::Array{T, 2})

Returns the Ideal Point of the nondominated frontier non_dom_sols.

compute_ideal_point(non_dom_sols::Union{Vector{MOPSolution}, Vector{BOPSolution}})

Returns the Ideal Point of the nondominated frontier non_dom_sols.

Modof.compute_nadir_point — Function.

compute_nadir_point{T<:Number}(non_dom_sols::Array{T, 2})

Returns the Nadir Point of the nondominated frontier non_dom_sols.

compute_nadir_point(non_dom_sols::Union{Vector{MOPSolution}, Vector{BOPSolution}})

Returns the Nadir Point of the nondominated frontier non_dom_sols.

Selecting points in a nondominated frontier

Modof.compute_closest_point_to_the_ideal_point — Function.

compute_closest_point_to_the_ideal_point{T<:Number}(non_dom_sols::Array{T, 2}, ideal_point::Vector{T}, return_pos::Bool=false)

Returns the closest point in the nondominated frontier non_dom_sols from its Ideal Point ideal_point.

compute_closest_point_to_the_ideal_point{T<:Number}(non_dom_sols::Union{Vector{MOPSolution}, Vector{BOPSolution}}, ideal_point::Vector{T})

Returns the closest point in the nondominated frontier non_dom_sols from its Ideal Point ideal_point.

compute_closest_point_to_the_ideal_point{T<:Number}(non_dom_sols::Array{T, 2})

Returns the closest point in the nondominated frontier non_dom_sols from its Ideal Point.

compute_closest_point_to_the_ideal_point(non_dom_sols::Union{Vector{MOPSolution}, Vector{BOPSolution}})

Returns the closest point in the nondominated frontier non_dom_sols from its Ideal Point.

Modof.compute_farthest_point_to_the_nadir_point — Function.

compute_farthest_point_to_the_nadir_point{T<:Number}(non_dom_sols::Array{T, 2}, nadir_point::Vector{T}, return_pos::Bool=false)

Returns the farthest point in the nondominated frontier non_dom_sols from its Nadir Point nadir_point.

compute_farthest_point_to_the_nadir_point{T<:Number}(non_dom_sols::Union{Vector{MOPSolution}, Vector{BOPSolution}}, nadir_point::Vector{T})

Returns the farthest point in the nondominated frontier non_dom_sols from its Nadir Point nadir_point.

compute_farthest_point_to_the_nadir_point{T<:Number}(non_dom_sols::Array{T, 2})

Returns the farthest point in the nondominated frontier non_dom_sols from its Nadir Point.

compute_farthest_point_to_the_nadir_point(non_dom_sols::Union{Vector{MOPSolution}, Vector{BOPSolution}})

Returns the farthest point in the nondominated frontier non_dom_sols from its Nadir Point.