Types

Type Definitions:

Storing instances of different classes of optimization problems

Modof.MOLPInstance — Type.

MOLPInstance

Datatype to store a Multi Objective Linear Program

Modof.MOBPInstance — Type.

MOBPInstance

Datatype to store a Multi Objective Binary Linear Program

Modof.MOIPInstance — Type.

MOIPInstance

Datatype to store a Multi Objective Integer Linear Program

Modof.MOMBLPInstance — Type.

MOMBLPInstance

Datatype to store a Multi Objective Mixed Binary Linear Program

Modof.MOMILPInstance — Type.

MOMILPInstance

Datatype to store a Multi Objective Mixed Integer Linear Program

Modof.BOLPInstance — Type.

BOLPInstance

Datatype to store a Bi-Objective Linear Program

Modof.BOBPInstance — Type.

BOBPInstance

Datatype to store a Bi-Objective Binary Linear Program

Modof.BOIPInstance — Type.

BOIPInstance

Datatype to store a Bi-Objective Integer Linear Program

Modof.BOMBLPInstance — Type.

BOMBLPInstance

Datatype to store a Bi-Objective Mixed Binary Linear Program

Modof.BOMILPInstance — Type.

BOMILPInstance

Datatype to store a Bi-Objective Mixed Integer Linear Program

Modof.OOESInstance — Type.

OOESInstance

Datatype to store an instance corresponding to minimizing a linear function over the efficient set

Storing solutions of different classes of optimization problems

Modof.MOPSolution — Type.

MOPSolution

Datatype to store a Multi Objective Optimization Solution

Modof.BOPSolution — Type.

BOPSolution

Datatype to store a Bi-Objective Optimization Solution

Modof.BOMSolution — Type.

BOMSolution

Datatype to store a Bi-Objective Mixed Optimization Solution

Modof.OOESSolution — Type.

OOESSolution

Datatype to store the solution of minimizing a linear function over the efficient set